Refactored so that policies are a seperate stack from the firehose creation process
